+++ Window/Module/chilledbump.h
@@ -27,8 +27,8 @@ public:
 
     SwitchCmd     m_eSwitchCmd;        //启停命令
 
-    uint8_t       m_usFreqSet;         //频率设置
-    uint8_t       m_usFreq;            //频率反馈
+    uint16_t       m_usFreqSet;         //频率设置
+    uint16_t       m_usFreq;            //频率反馈
 
     bool          m_xRemote;           //远程/本地
     bool          m_xRunningFlag;      //运行标志
